feat(ds): promote AtomicChallengeCard + rail to design system
Promote the Widgetbook-prototyped atomic challenge card primitives
(AtomicChallengeCard, AtomicChallengeRail, OngoingRailFrame + phase/rail
treatment enums) into lib/design_system/src so the new Fair Rewards
challenges surface can consume them as stable DS components.

- src/atomic_challenge_card.dart (Material container, not banned M3 Card)
- test/design_system/atomic_challenge_card_test.dart (phases, fill clamp,
  null state-only rail, checkbox/technicalOngoing treatments, tap)
- barrel export + DESIGN_SYSTEM.md catalog row + AtomicChallengeCard.genesis.md
  (records the gap-proof approval trail for the new pattern)
- stories (card/detail/bands/profile) now import the widget from the DS
  instead of defining it inline; widgetbook codegen regenerated

verify-widget.sh AtomicChallengeCard: 11/11 passed.

Relates to #449
