+++ widgetbook/lib/stories/active_challenge_bands.stories.dart
@@ -4,8 +4,6 @@ import 'package:widgetbook/widgetbook.dart';
 
 import 'package:crypto_mobile_app/design_system/design_system.dart';
 
-import 'atomic_challenge_card.stories.dart';
-
 part 'active_challenge_bands.stories.g.dart';
 
 const meta = Meta<ActiveChallengeBands>(path: 'prototypes/challenges');
